requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
psycopg==3.1.18
python-dotenv==1.0.1
typing_extensions>=4.12.2
passlib[bcrypt]
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests
reportlab
qrcode[pil]
python-barcode[images]
